requirements.txt
supabase-py
streamlit
gspread
oauth2client
pandas
numpy
openai